# Conflicts:
#	src/main/java/group/depapp/config/AppConfig.java

+++ src/main/java/group/depapp/util/Synchronizer.java
@@ -1,18 +1,31 @@
 package group.depapp.util;
 
 import group.depapp.domain.DepartmentDTO;
+import group.depapp.service.DepartmentService;
 import group.depapp.service.DepartmentServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+@Service
 public class Synchronizer {
 
-    public void synchronize(){
-        XMLHandler xmlHandler = new XMLHandler();
-        DepartmentServiceImpl departmentService = new DepartmentServiceImpl();
+    DepartmentService departmentService;
+
+    @Autowired
+    public Synchronizer(DepartmentService departmentService) {
+        this.departmentService = departmentService;
+    }
+
+    @Transactional
+    public void synchronize() {
+        XMLHandler xmlHandler = new XMLHandler(departmentService);
+
 
         List<DepartmentDTO> dtosFromXML = xmlHandler.parse();
         List<DepartmentDTO> dtosFromDB = departmentService.getAll();
@@ -37,7 +50,7 @@ public class Synchronizer {
                         if (!entry.getKey().getDescription().equals(departmentDTO.getDescription())) {
                             departmentService.update(entry.getKey());
                         }
-                     entry.setValue(true);
+                        entry.setValue(true);
 
                     }
                 }
@@ -47,8 +60,9 @@ public class Synchronizer {
         for (Map.Entry<DepartmentDTO, Boolean> entry : mapForSynchronization.entrySet()) {
             if (!entry.getValue()) departmentsToInsert.add(entry.getKey());
         }
-
-        departmentService.save(departmentsToInsert);
-        departmentService.delete(departmentsToDelete);
+        if (departmentsToInsert.size() > 0)
+            departmentService.save(departmentsToInsert);
+        if (departmentsToDelete.size() > 0)
+            departmentService.delete(departmentsToDelete);
     }
 }
